BizBackend/requirements.txt
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
billiard==4.2.1
celery==5.4.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
Django==5.1.7
django-admin==2.0.2
django-cors-headers==4.7.0
django-excel-response2==3.0.6
django-filter==25.1
django-restframework==0.0.1
django-six==1.0.5
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
excel-base==1.0.4
gunicorn==23.0.0
idna==3.10
isoweek==1.3.3
kombu==5.5.1
numpy==2.2.4
packaging==24.2
pandas==2.2.3
prompt_toolkit==3.0.50
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
redis==5.2.1
requests==2.32.3
screen==1.0.1
six==1.17.0
sqlparse==0.5.3
TimeConvert==3.0.13
typing_extensions==4.13.0
tzdata==2025.1
tzlocal==5.3.1
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
whitenoise==6.9.0
xlwt==1.3.0
